requirements.txt
# FastAPI & Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Pydantic & Validation
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# LangChain RAG
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1

# Embeddings
sentence-transformers==2.7.0
openai==1.30.1  # optional, for OpenAI embeddings

# Vector Database
chromadb==0.5.0
faiss-cpu==1.8.0

# Reranking
sentence-transformers==2.7.0  # cross-encoder included

# Document Processing
pypdf2==3.0.1
pdfplumber==0.11.1
python-docx==1.1.0
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
tenacity==8.3.0
loguru==0.7.2
numpy==1.26.4
tiktoken==0.7.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# Dev Tools
ruff==0.4.4
